JAMES-3028 move S3BlobConfiguration to blob-s3 and rename beans
